replace get_subcell_topo by bezier_nodedup
This patch replaces the get_subcell_topo method by an alternative sampling
scheme in the form of bezier_nodedup, which can be used to draw child elements
via the sample's hull property. The benefit of this approach is that it
continues to work in nutils' master branch which has repurposed the 'simplices'
attribute. Additionally, the surface plot is replaced by an outline of the
trimmed domain's boundary.