replace elems argument by vpe (voxels per element)
This patch removes the elems argument, which took two integers, by a the vpe
argument, taking a single float. Vpe stands for voxels per element and is used
to scale the image coordinates to topological coordinates under the assumption
that the aspect ratio of elements in the ambient domain should be approximately
square. This change makes it straightfoward to trim images of different aspect
ratios.